Move _OutsideWrapper to leaf module to avoid circular imports in PR 3
